Split 'Party-functions.js' into two files, one for each Party component
